Fix Navigator Key Navigation Explosion
